+++ kitty/state.h
@@ -65,7 +65,8 @@ typedef struct Options {
     float cursor_trail_decay_fast;
     float cursor_trail_decay_slow;
     color_type cursor_trail_color;
-    float cursor_trail_start_threshold;
+    int cursor_trail_start_threshold_x;
+    int cursor_trail_start_threshold_y;
     unsigned int url_style;
     unsigned int scrollback_pager_history_size;
     bool scrollback_fill_enlarged_window;